New collision detection and fix to Y ball collision
previously when the ball hit the paddle the Y velocity would switch
between -1 to 1. With this new change I use 6 different bounding boxes
to detect where the ball hit the paddle and change the Y velocity based
on that collision.

Each paddle has an over all bounding box that incompouses the entire
paddle. This large bounding box is used to make collision detection
faster than check each box every tick. After a collision is found
between the ball and a paddle the we find the section of the paddle that
the ball hit and change the Y velocity accordingly. The values are
below:

Top Corner – 2f
Top Center – 1f
Center – 0f
Bottom Center + 1f
Bottom Corner + 2f