Add Intro skips and Main Menu delay
I added a way to skip both intros (A button or esc key). I all ways
hated no being able to skip intro in games so I thought I should not be
a hypocrite. This caused the Main menu to be skipped due to the way
keyboard and game pad input is handled in XNA. So I added a 5000
millisecond pause before interaction is allowed.